Split classifier responsibilities
